pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codehist"
version = "0.1.0"
description = "Universal parser for AI coding agent chat histories and workspace data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "codervisor", email = "codervisor@example.com"}
]
keywords = ["ai", "coding", "chat", "history", "copilot", "cursor", "windsurf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Tools",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pathlib",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
analysis = [
    "pandas>=1.5.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
]
export = [
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "pygments>=2.15.0",
]

[project.urls]
Homepage = "https://github.com/codervisor/codehist"
Repository = "https://github.com/codervisor/codehist"
Issues = "https://github.com/codervisor/codehist/issues"
Documentation = "https://github.com/codervisor/codehist/docs"

[project.scripts]
codehist = "codehist.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["codehist*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=codehist --cov-report=term-missing"
